[MRG] Adapt to scikit-learn 1.0 + fix deprecated dtypes (#406)

+++ tslearn/utils/utils.py
@@ -146,14 +146,14 @@ def to_time_series(ts, remove_nans=False):
     ts_out = numpy.array(ts, copy=True)
     if ts_out.ndim <= 1:
         ts_out = ts_out.reshape((-1, 1))
-    if ts_out.dtype != numpy.float:
-        ts_out = ts_out.astype(numpy.float)
+    if ts_out.dtype != float:
+        ts_out = ts_out.astype(float)
     if remove_nans:
         ts_out = ts_out[:ts_size(ts_out)]
     return ts_out
 
 
-def to_time_series_dataset(dataset, dtype=numpy.float):
+def to_time_series_dataset(dataset, dtype=float):
     """Transforms a time series dataset so that it fits the format used in
     ``tslearn`` models.
 
@@ -162,7 +162,7 @@ def to_time_series_dataset(dataset, dtype=numpy.float):
     dataset : array-like
         The dataset of time series to be transformed. A single time series will
         be automatically wrapped into a dataset with a single entry.
-    dtype : data type (default: numpy.float)
+    dtype : data type (default: float)
         Data type for the returned dataset.
 
     Returns
@@ -682,7 +682,7 @@ def _load_arff_uea(dataset_path):
 
     else:  # uni-variate situation
         x_ = data[names[:-1]]
-        x = numpy.asarray(x_.tolist(), dtype=numpy.float32)
+        x = numpy.asarray(x_.tolist(), dtype=float)
         x = x.reshape(len(x), -1, 1)
 
     return x, y
@@ -710,5 +710,5 @@ def _load_txt_uea(dataset_path):
     """
     data = numpy.loadtxt(dataset_path)
     X = to_time_series_dataset(data[:, 1:])
-    y = data[:, 0].astype(numpy.int)
+    y = data[:, 0].astype(int)
     return X, y
